provision/configure: split files for maintainability.
