+++ provision/configure/bootstrap.go
@@ -8,25 +8,16 @@ import (
 	"context"
 	"errors"
 	"flag"
-	"io"
 	"io/fs"
-	"io/ioutil"
 	"log"
-	"net"
 	"os"
 
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 	"namespacelabs.dev/foundation/internal/engine/ops/defs"
 	"namespacelabs.dev/foundation/internal/fnerrors"
 	"namespacelabs.dev/foundation/internal/fnfs/memfs"
 	"namespacelabs.dev/foundation/internal/logoutput"
-	"namespacelabs.dev/foundation/internal/versions"
-	"namespacelabs.dev/foundation/provision/tool/grpcstdio"
 	"namespacelabs.dev/foundation/provision/tool/protocol"
-	"namespacelabs.dev/foundation/schema"
 )
 
 func init() {
@@ -36,31 +27,6 @@ func init() {
 type Request struct {
 	Snapshots map[string]fs.FS
 	r         *protocol.ToolRequest
-}
-
-type StackRequest struct {
-	Request
-	Env   *schema.Environment
-	Focus *schema.Stack_Entry
-	Stack *schema.Stack
-}
-
-type MakeExtension interface {
-	ToDefinition() (*schema.DefExtension, error)
-}
-
-type ApplyOutput struct {
-	Definitions []defs.MakeDefinition
-	Extensions  []MakeExtension
-}
-
-type DeleteOutput struct {
-	Ops []defs.MakeDefinition
-}
-
-type StackHandler interface {
-	Apply(context.Context, StackRequest, *ApplyOutput) error
-	Delete(context.Context, StackRequest, *DeleteOutput) error
 }
 
 // XXX remove Tool when all the uses are gone.
@@ -85,16 +51,12 @@ func (p Request) UnpackInput(msg proto.Message) error {
 		}
 	}
 
-	return fnerrors.InternalError("no such env: %s", msg.ProtoReflect().Descriptor().FullName())
+	return fnerrors.InternalError("no such input: %s", msg.ProtoReflect().Descriptor().FullName())
 }
 
 // PackageOwner returns the name of the package that defined this tool.
 func (p Request) PackageOwner() string {
 	return p.r.GetToolPackage()
-}
-
-func RunTool(t Tool) {
-	run(context.Background(), handlerCompat{t})
 }
 
 func run(ctx context.Context, h AllHandlers) {
@@ -107,83 +69,6 @@ func run(ctx context.Context, h AllHandlers) {
 	}
 }
 
-func handle(ctx context.Context, h AllHandlers) error {
-	conn, err := grpc.DialContext(ctx, "stdio",
-		grpc.WithInsecure(),
-		grpc.WithReadBufferSize(0),
-		grpc.WithWriteBufferSize(0),
-		grpc.WithContextDialer(func(_ context.Context, _ string) (net.Conn, error) {
-			return grpcstdio.NewConnection(os.Stdout, os.Stdin), nil
-		}))
-	if err != nil {
-		return err
-	}
-
-	defer conn.Close()
-
-	cli := protocol.NewInvocationServiceClient(conn)
-	stream, err := cli.Worker(ctx)
-	if err != nil {
-		return err
-	}
-
-	if err := stream.Send(&protocol.WorkerChunk{ClientHello: &protocol.WorkerChunk_ClientHello{
-		FnApiVersion:   versions.APIVersion,
-		ToolApiVersion: versions.ToolAPIVersion,
-	}}); err != nil {
-		return err
-	}
-
-	for {
-		msg, err := stream.Recv()
-		if err != nil {
-			return err
-		}
-
-		if msg.ToolRequest != nil {
-			response, err := handleRequest(ctx, msg.ToolRequest, h)
-			if err != nil {
-				return err
-			}
-
-			if err := stream.Send(&protocol.WorkerChunk{ToolResponse: response}); err != nil {
-				return err
-			}
-
-			if err := stream.CloseSend(); err != nil {
-				return err
-			}
-
-			// Make sure that the send was received.
-			if _, err := stream.Recv(); err != nil {
-				if err != io.EOF {
-					return err
-				}
-			}
-
-			return nil
-		}
-	}
-}
-
-type handlerCompat struct {
-	tool Tool
-}
-
-var _ AllHandlers = handlerCompat{}
-
-func (h handlerCompat) Apply(ctx context.Context, req StackRequest, output *ApplyOutput) error {
-	return h.tool.Apply(ctx, req, output)
-}
-
-func (h handlerCompat) Delete(ctx context.Context, req StackRequest, output *DeleteOutput) error {
-	return h.tool.Delete(ctx, req, output)
-}
-
-func (h handlerCompat) Invoke(context.Context, Request) (*protocol.InvokeResponse, error) {
-	return nil, status.Error(codes.Unavailable, "invoke not supported")
-}
-
 func Handle(h *Handlers) {
 	run(context.Background(), runHandlers{h})
 }
@@ -192,31 +77,6 @@ func HandleInvoke(f InvokeFunc) {
 	h := NewHandlers()
 	h.Any().HandleInvoke(f)
 	Handle(h)
-}
-
-func runTool(ctx context.Context, r io.Reader, w io.Writer, t AllHandlers) error {
-	reqBytes, err := ioutil.ReadAll(r)
-	if err != nil {
-		return err
-	}
-
-	req := &protocol.ToolRequest{}
-	if err := proto.Unmarshal(reqBytes, req); err != nil {
-		return err
-	}
-
-	response, err := handleRequest(ctx, req, t)
-	if err != nil {
-		return err
-	}
-
-	serialized, err := proto.Marshal(response)
-	if err != nil {
-		return err
-	}
-
-	w.Write(serialized)
-	return nil
 }
 
 func handleRequest(ctx context.Context, req *protocol.ToolRequest, t AllHandlers) (*protocol.ToolResponse, error) {
@@ -291,29 +151,4 @@ func handleRequest(ctx context.Context, req *protocol.ToolRequest, t AllHandlers
 	}
 
 	return response, nil
-}
-
-func parseStackRequest(br Request, header *protocol.StackRelated) (StackRequest, error) {
-	if header == nil {
-		// This is temporary, while we move from top-level fields to {Apply,Delete} specific ones.
-		header = &protocol.StackRelated{
-			FocusedServer: br.r.FocusedServer,
-			Env:           br.r.Env,
-			Stack:         br.r.Stack,
-		}
-	}
-
-	var p StackRequest
-
-	s := header.Stack.GetServer(schema.PackageName(header.FocusedServer))
-	if s == nil {
-		return p, fnerrors.InternalError("%s: focused server not present in the stack", header.FocusedServer)
-	}
-
-	p.Request = br
-	p.Env = header.Env
-	p.Focus = s
-	p.Stack = header.Stack
-
-	return p, nil
 }
